requirements.txt
streamlit
pandas
openpyxl
tabulate
python-dotenv
crewai
google-genai